pyproject.toml
[project]
name = "new-rag-project-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.7.4",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "langchain-tavily>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "streamlit>=1.38.0",
    "watchdog>=6.0.0",
]
